requirements.txt
# Core dependencies
google-adk>=1.0.0
google-cloud-firestore>=2.0.0
google-genai
uvicorn
fastapi
python-telegram-bot>=20.0
firebase-admin>=6.0.0
slowapi>=0.1.9
python-json-logger>=2.0.7
prometheus-client>=0.20.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
